Turn rest of generator modules into packages
